weird logjam in scheduler -- as if the select()  would not get read flag when child is dead and reaped, but the socket/pipe is still still very much around ??  Perhaps Linux kernel 2.4.5-pre4 is the culprit ??

+++ scheduler/conf.c
@@ -40,7 +40,7 @@ int	default_full_content = 1;
 
 /* If we use ZMailer malloc library in debug mode, BAD-PTR looks like
    following value.. */
-void *BADPTR = (void*) 0x5555555555555555L;
+void *BADPTR = (void*) (unsigned long)0x5555555555555555L;
 
 /* Push reports regarding recipients out this often, unless
    parametrized to do differently */
